logic: extract AddressBookSession class
In the future, we want to expose the lastShownList as an ObservableList,
so we can display it as a JavaFX TableView.

Furthermore, to provide more direct user feedback, we also want the
lastShownList to be kept in sync with the operations on the address
book. For example, if a person is deleted from the address book, it
should also be deleted from the lastShownList as well. Likewise, if a
person is added, it should also be added to the lastShownList.

However, currently the AddressBook and lastShownList are passed directly
to the Commands. This is dangerous, as it does not guarantee that the
lastShownList will be kept in sync with the AddressBook. For example, a
rogue command may delete a person from the address book, but not delete
that person from the lastShownList as well. This will confuse the user.

As such, introduce a new AddressBookSession class, which will manage
both the AddressBook and lastShownList. In the future, all operations on
the AddressBook (e.g. addPerson, removePerson etc.) will go through this
class, and this class will ensure that the lastShownList is kept in
sync.

For now, we introduce the `addressBookSession` member in the base
Command class, while keeping the `addressBook` and `relevantPersons`
members to keep existing code working. In the upcoming commits, we will
fully switch all commands to only use `addressBookSession`.

At the same time, we love tests, so introduce an AddressBookSessionTest
suite that is dedicated to testing the AddressBookSession class.