commands: extract containsPerson() into AddressBookSession
ViewAllCommand and ViewCommand call addressBook.containsPerson()
directly to determine if the address book contains a person.

To make the API pleasant to use, we want to make sure that the
addressBook is fully encapsulated within the addressBookSession.

As such, extract out the "addressBook.containsPerson()" functionality
into addressBookSession.containsPerson(), and make the above classes
call that instead.