commands: extract {get,set}Filter() into AddressBookSession
FindCommand and ListCommand currently filters and manages the
lastShownList entries by accessing the addressBook directly.

We want lastShownList and addressBook to be fully encapsulated within
AddressBookSession. At the same time, we need to provide
AddressBookSession with the knowledge of whether there is a filter, and
what is the filter function. This is useful, for example, when the list
is filtered with the "find" command, and then a new person is added.
Whether the person is added to the lastShownList as well would depend on
whether the person matches the filter or not, but this also requires the
AddressBookSession to know what filter is currently used.

As such, extract the filtering functionality into the
AddressBookSession. Rather than doing the filtering manually, make
FindCommand and ListCommand call addressBookSession.setFilter() to do
the filtering for them.