AddressBookSession.addPerson(): display added person in lastShownList
To make the UI easier to use, the lastShownList should be synchronized
with operations on the address book.

However, currently when adding a new person, the new person does not
show up on the addressBook.

Let's fix that by adding the new person to the lastShownList as well if
it passes the current filter, if any.